тесты 74% 94% 100% соответственно
