+++ tests/test_generators.py
@@ -1,6 +1,6 @@
 import pytest
 from typing import Dict, Any, List
-from src.generators import filter_by_currency
+from src.generators import filter_by_currency, transaction_descriptions, card_number_generator
 
 
 def test_filter_bay_currency() -> None:
@@ -29,126 +29,246 @@ def test_filter_bay_currency() -> None:
 
 
 @pytest.fixture
-def transactions_fixture() -> List[Dict[str, Any]]:
-    """Фикстура с тестовыми транзакциями."""
+def sample_transactions() -> List[Dict[str, Any]]:
+    """Фикстура с различными типами транзакций для тестирования"""
     return [
-        # Корректные транзакции
-        {"id": 1, "operationAmount": {"amount": "100.00", "currency": {"code": "USD", "name": "Доллар"}}},
-        {"id": 2, "operationAmount": {"amount": "200.00", "currency": {"code": "EUR", "name": "Евро"}}},
-        {"id": 3, "operationAmount": {"amount": "300.00", "currency": {"code": "USD", "name": "Доллар"}}},
-        # Некорректные/неполные транзакции
-        {"id": 4, "operationAmount": {"amount": "400.00"}},  # Нет currency
-        {"id": 5},  # Нет operationAmount
+        # Стандартные случаи
+        {"id": 1, "operationAmount": {"currency": {"code": "USD"}}},
+        {"id": 2, "operationAmount": {"currency": {"code": "EUR"}}},
+        {"id": 3, "operationAmount": {"currency": {"code": "USD", "name": "Доллар"}}},
+
+        # Крайние случаи
+        {"id": 4},  # Нет operationAmount
+        {"id": 5, "operationAmount": None},  # operationAmount = None
+        {"id": 6, "operationAmount": {"amount": "100"}},  # Нет currency
+        {"id": 7, "operationAmount": {"currency": None}},  # currency = None
+        {"id": 8, "operationAmount": {"currency": {"name": "Доллар"}}},  # Нет code
+        {"id": 9, "operationAmount": {"currency": "USD"}},  # currency как строка
+        {"id": 10, "operationAmount": {"currency": {"code": 840}}},  # code как число
     ]
 
 
-@pytest.mark.parametrize(
-    "currency, expected_ids",
-    [
-        ("USD", [1, 3]),  # Ожидаются транзакции с ID 1 и 3
-        ("EUR", [2]),  # Ожидается транзакция с ID 2
-        ("RUB", []),  # Нет подходящих транзакций
-        ("GBP", []),  # Нет подходящих транзакций
-    ],
-)
-def test_filter_by_currency(
-    transactions_fixture: List[Dict[str, Any]],
-    currency: str,
-    expected_ids: List[int],
+@pytest.mark.parametrize("currency, expected_ids", [
+    ("USD", [1, 3]),  # Стандартные USD транзакции
+    ("EUR", [2]),  # Стандартные EUR транзакции
+    ("RUB", []),  # Нет RUB транзакций
+    ("GBP", []),  # Нет GBP транзакций
+])
+def test_standard_cases(
+        sample_transactions: List[Dict[str, Any]],
+        currency: str,
+        expected_ids: List[int]
 ) -> None:
-    """Тест фильтрации транзакций по валюте."""
-    result = list(filter_by_currency(transactions_fixture, currency))
-    assert len(result) == len(expected_ids)
-    assert all(tx["id"] in expected_ids for tx in result)
+    """Тестирование стандартных случаев фильтрации"""
+    result = list(filter_by_currency(sample_transactions, currency))
+    assert [tx["id"] for tx in result] == expected_ids
 
 
-def test_filter_by_currency_empty_input() -> None:
-    """Тест на пустом списке транзакций."""
+def test_edge_cases(sample_transactions: List[Dict[str, Any]]) -> None:
+    """Тестирование крайних случаев"""
+    # Все некорректные транзакции должны быть проигнорированы
+    assert len(list(filter_by_currency(sample_transactions, "USD"))) == 2
+    assert len(list(filter_by_currency(sample_transactions, "EUR"))) == 1
+
+
+def test_empty_input() -> None:
+    """Тестирование пустого ввода"""
     assert list(filter_by_currency([], "USD")) == []
 
 
-def test_filter_by_currency_invalid_structure() -> None:
-    """Тест с транзакциями, у которых нет поля currency."""
-    transactions = [
-        {"id": 1, "operationAmount": {"amount": "100.00"}},
-        {"id": 2, "operationAmount": {"amount": "200.00", "currency": {"code": "USD"}}},
-    ]
-    assert list(filter_by_currency(transactions, "USD")) == [transactions[1]]
-
-
-def test_filter_multiple_currencies() -> None:
-    """Тестирование фильтрации при смешанных валютах."""
-    transactions = [
-        {"operationAmount": {"currency": {"code": "USD"}}},
-        {"operationAmount": {"currency": {"code": "EUR"}}},
-        {"operationAmount": {"currency": {"code": "USD"}}},
-        {"operationAmount": {"amount": "100.00"}},  # Нет currency
-    ]
+def test_currency_case_sensitivity() -> None:
+    """Тестирование чувствительности к регистру"""
+    transactions = [{"operationAmount": {"currency": {"code": "usd"}}}]
+    # В зависимости от требований - должен ли код валюты быть в верхнем регистре
     result = list(filter_by_currency(transactions, "USD"))
-    assert len(result) == 2
-    assert all(tx["operationAmount"]["currency"]["code"] == "USD" for tx in result)
+    assert len(result) == 0  # или 1, если функция нечувствительна к регистру
 
 
-@pytest.mark.parametrize("transaction,currency,expected", [
-    # Полная корректная структура
-    (
-        {
-            "operationAmount": {
-                "currency": {"code": "USD", "name": "Доллар"}
-            }
-        },
-        "USD",
-        True
-    ),
-    # Нет поля currency
-    (
-        {"operationAmount": {"amount": "100.00"}},
-        "USD",
-        False
-    ),
-    # Нет operationAmount
-    (
-        {"id": 1},
-        "USD",
-        False
-    ),
-    # Currency как строка вместо словаря
-    (
-        {"operationAmount": {"currency": "USD"}},
-        "USD",
-        False
-    ),
-    # Пустой словарь currency
-    (
-        {"operationAmount": {"currency": {}}},
-        "USD",
-        False
-    ),
-    # Несколько уровней вложенности
-    (
-        {"data": {"operationAmount": {"currency": {"code": "EUR"}}}},
-        "EUR",
-        False  # Не будет найдено из-за неправильного пути
-    ),
-])
-def test_filter_edge_cases(transaction: Dict[str, Any], currency: str, expected: bool):
-    """Тестирование различных структур транзакций."""
-    result = list(filter_by_currency([transaction], currency))
-    assert (len(result) == 1) == expected
+def test_nested_transaction_structures() -> None:
+    """Тестирование вложенных структур"""
+    transactions = [
+        {"data": {"operation": {"amount": {"currency": {"code": "USD"}}}}}
+    ]
+    # В зависимости от реализации функции
+    result = list(filter_by_currency(transactions, "USD"))
+    assert len(result) == 0  # или 1, если функция поддерживает глубокий поиск
 
 
-def test_filter_invalid_input_types() -> None:
-    """Тестирование обработки некорректных типов входных данных."""
-    # 1. Проверка неверного типа transactions
-    with pytest.raises(TypeError, match="transactions должен быть списком"):
-        list(filter_by_currency("not a list", "USD"))
+def test_performance_large_dataset() -> None:
+    """Тестирование производительности на большом наборе данных"""
+    large_transactions = [{"operationAmount": {"currency": {"code": "USD"}}}] * 10000
+    result = list(filter_by_currency(large_transactions, "USD"))
+    assert len(result) == 10000
 
-    # 2. Проверка неверного типа currency
-    with pytest.raises(TypeError, match="currency должен быть строкой из 3 символов"):
-        list(filter_by_currency([], 123))
 
-    # 3. Проверка неверной длины currency
-    with pytest.raises(TypeError, match="currency должен быть строкой из 3 символов"):
-        list(filter_by_currency([], "US"))  # 2 символа
-    with pytest.raises(TypeError):
-        list(filter_by_currency([], "USDX"))  # 4 символа
+@pytest.fixture
+def transactions_fixture() -> List[Dict[str, Any]]:
+    """Фикстура с тестовыми транзакциями (корректные и некорректные)."""
+    return [
+        {"id": 1, "description": "Перевод организации"},
+        {"id": 2, "description": "Покупка валюты"},
+        {"id": 3},  # Нет описания
+        {"id": 4, "description": None},  # Описание = None
+        {"id": 5, "description": 12345},  # Не строка
+        {"id": 6, "description": "Снятие наличных"},
+    ]
+
+
+def test_transaction_descriptions_valid(transactions_fixture: List[Dict[str, Any]]) -> None:
+    """Тест корректных описаний (только строки)."""
+    generator = transaction_descriptions(transactions_fixture)
+    expected_descriptions = ["Перевод организации", "Покупка валюты", "Снятие наличных"]
+    assert list(generator) == expected_descriptions
+
+
+def test_transaction_descriptions_empty_list() -> None:
+    """Тест пустого списка транзакций."""
+    generator = transaction_descriptions([])
+    assert list(generator) == []
+
+
+def test_transaction_descriptions_no_description(transactions_fixture: List[Dict[str, Any]]) -> None:
+    """Тест транзакций без описания (пропускаются)."""
+    # Выбираем только транзакции без description или с некорректным типом
+    invalid_transactions = [tx for tx in transactions_fixture if not isinstance(tx.get("description"), str)]
+    generator = transaction_descriptions(invalid_transactions)
+    assert list(generator) == []
+
+
+@pytest.mark.parametrize(
+    "description, expected",
+    [
+        ("Оплата налогов", ["Оплата налогов"]),  # Корректная строка
+        (None, []),  # None пропускается
+        (123, []),  # Не строка
+        ("", [""]),  # Пустая строка (допустимо)
+    ],
+)
+def test_transaction_descriptions_parametrized(description: Any, expected: List[str]) -> None:
+    """Параметризованный тест для разных типов description."""
+    transactions = [{"id": 1, "description": description}]
+    generator = transaction_descriptions(transactions)
+    assert list(generator) == expected
+
+
+@pytest.mark.parametrize(
+    "transactions, expected",
+    [
+        ([{"description": "Test"}], ["Test"]),
+        ([{"description": "A"}, {"description": "B"}], ["A", "B"]),
+        ([{}], []),
+    ],
+)
+def test_transaction_descriptions_cases(transactions: Any, expected: List[str]) -> None:
+    assert list(transaction_descriptions(transactions)) == expected
+
+
+def test_long_description() -> None:
+    long_text = "A" * 1000  # Very long description
+    transactions = [{"description": long_text}]
+    result = list(transaction_descriptions(transactions))
+    assert result == [long_text], "Должен поддерживать длинные описания"
+
+
+def test_card_number_generator() -> None:
+    # Тест 1: Генерация одного номера
+    gen = card_number_generator(1, 1)
+    assert next(gen) == "0000 0000 0000 0001", "Первый номер должен быть 0000 0000 0000 0001"
+
+    # Тест 2: Генерация диапазона
+    gen = card_number_generator(1, 3)
+    assert list(gen) == [
+        "0000 0000 0000 0001",
+        "0000 0000 0000 0002",
+        "0000 0000 0000 0003",
+    ], "Должны сгенерироваться три последовательных номера"
+
+    # Тест 3: Проверка формата
+    gen = card_number_generator(1234567812345678, 1234567812345678)
+    number = next(gen)
+    assert len(number) == 19, "Длина номера должна быть 19 символов (16 цифр + 3 пробела)"
+    assert number.count(" ") == 3, "Должно быть 3 пробела"
+    assert number.replace(" ", "").isdigit(), "Должны быть только цифры"
+
+    # Тест 4: Некорректный диапазон (start > end)
+    try:
+        list(card_number_generator(10, 1))
+        assert False, "Должно было возникнуть ValueError"
+    except ValueError:
+        pass  # Ожидаемое поведение
+
+    # Тест 5: Граничные значения
+    gen = card_number_generator(9999999999999999, 9999999999999999)
+    assert next(gen) == "9999 9999 9999 9999", "Последний номер должен быть 9999 9999 9999 9999"
+
+
+@pytest.mark.parametrize(
+    "start, end, expected_numbers",
+    [
+        (
+            1,
+            3,
+            [
+                "0000 0000 0000 0001",
+                "0000 0000 0000 0002",
+                "0000 0000 0000 0003",
+            ],
+        ),
+        (
+            9999,
+            10001,
+            [
+                "0000 0000 0000 9999",
+                "0000 0000 0001 0000",
+                "0000 0000 0001 0001",
+            ],
+        ),
+    ],
+)
+def test_card_number_generator_range(
+    start: int,
+    end: int,
+    expected_numbers: list[str],
+) -> None:
+    """Тест генерации номеров карт в заданном диапазоне."""
+    generator = card_number_generator(start, end)
+    generated_numbers = list(generator)
+    assert generated_numbers == expected_numbers
+
+
+def test_card_number_generator_format() -> None:
+    """Тест формата номера карты (16 цифр, разделённых пробелами)."""
+    generator = card_number_generator(1234567890123456, 1234567890123456)
+    number = next(generator)
+    assert len(number) == 19  # 16 цифр + 3 пробела
+    assert number.replace(" ", "").isdigit()
+
+
+def test_card_number_generator_invalid_range() -> None:
+    """Тест на некорректный диапазон (start > end)."""
+    with pytest.raises(ValueError, match="Start must be less than or equal to end"):
+        list(card_number_generator(10, 1))
+
+
+def test_card_number_generator_edge_cases() -> None:
+    """Тест граничных случаев (минимальный и максимальный номера)."""
+    # Первый номер
+    assert next(card_number_generator(1, 1)) == "0000 0000 0000 0001"
+    # Последний номер
+    assert next(card_number_generator(9999999999999999, 9999999999999999)) == "9999 9999 9999 9999"
+
+
+def test_large_range() -> None:
+    gen = card_number_generator(1, 100)
+    result = list(gen)
+    assert len(result) == 100, "Должен генерировать большой диапазон"
+    assert result[0] == "0000 0000 0000 0001", "Первый номер должен быть корректным"
+    assert result[-1] == "0000 0000 0000 0100", "Последний номер должен быть корректным"
+
+
+def test_spacing_in_numbers() -> None:
+    gen = card_number_generator(1234567812345678, 1234567812345678)
+    number = next(gen)
+    parts = number.split(" ")
+    assert len(parts) == 4, "Должно быть 4 группы цифр"
+    assert all(len(part) == 4 for part in parts), "Каждая группа должна содержать 4 цифры"
